Search(): Use `feedkeys` instead of `normal!` to advance search
Because of `:h function-search-undo`, using `normal! n` at the end of
the function doesn't highlight the matches with 'hlsearch' (though it
does advance to the next match).

I've seen other mappings that add a `/<CR>` or `n` to the end of the
mapping, but that doesn't work here because it's an operator, so `g@`
needs to be the last thing in the mapping. And I was reluctant to work
out how to make it an `<expr>` mapping.

Fortunately, search highlighting works when `n` is used with
`feedkeys()`.